requirements.txt
psycopg2-binary
psycopg2
requests
load_dotenv
sqlalchemy
alembic
